Move parser spec types into cli_commands.py and drop ParserContext
`cli_parser.py` now imports `OptSpec` / `PositionalSpec` / `MutexGroup`
/ `SubcommandSpec` / `CommandSpec` / `ParsedCmd` from
`cli_commands.py` alongside the catalog itself. The dependency is
strictly one-way (`cli_parser → cli_commands`), so the helpers can go
back to referencing `COMMON_OPTIONS` and `COMMAND_BY_NAME_OR_ALIAS` as
module-level names and `parse_cmdline(argv)` regains its single-arg
signature - the short-lived `ParserContext` indirection is no longer
needed.
